Move changesetRewirer into rewirer.go
Since I want to be using the rewirer in the resolver layer as well, I don't think it strictly belongs into the service layer anymore and it should move away from the service_apply_campaign file.
